requirements.txt
streamlit
pytesseract
pdf2image
pillow
pandas
python-dateutil
